Work on accessibility
My mouse is broken, so I have to use keyboard navigation anyways, making
this the perfect time to investigate accessibility.

This first step is adding ARIA roles and attributes to many elements,
mainly the fake meters, but also using aria-label for the inputs that
don't have labels. I also added scope="col" to the table headers.

Edge is still complaining that most of the <select>s don't have labels,
and I need to completely redo the info tooltips. However, I still feel
that this is a good first step for the time being.